+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -8,6 +8,7 @@ import { groupController } from "~/modules/groups/groups.js";
 import { permissionController } from "~/modules/permissions/permissions.js";
 import { projectApiKeyController } from "~/modules/project-api-keys/project-api-keys.js";
 import { projectGroupController } from "~/modules/project-groups/project-groups.js";
+import { projectPermissionsController } from "~/modules/project-permissions/project-permissions.js";
 import { projectController } from "~/modules/projects/projects.js";
 import { userController, userService } from "~/modules/users/users.js";
 
@@ -23,6 +24,7 @@ const apiV1 = new BaseServerApplicationApi(
 	...authController.routes,
 	...permissionController.routes,
 	...projectGroupController.routes,
+	...projectPermissionsController.routes,
 	...projectController.routes,
 	...contributorController.routes,
 	...userController.routes,