+++ apps/frontend/src/libs/modules/store/libs/types/extra-arguments.type.ts
@@ -5,6 +5,8 @@ import { type contributorApi } from "~/modules/contributors/contributors.js";
 import { type groupApi } from "~/modules/groups/groups.js";
 import { type permissionApi } from "~/modules/permissions/permissions.js";
 import { type projectApiKeysApi } from "~/modules/project-api-keys/project-api-keys.js";
+import { type projectGroupApi } from "~/modules/project-groups/project-groups.js";
+import { type projectPermissionsApi } from "~/modules/project-permissions/project-permissions.js";
 import { type projectApi } from "~/modules/projects/projects.js";
 import { type userApi } from "~/modules/users/users.js";
 
@@ -15,6 +17,8 @@ type ExtraArguments = {
 	permissionApi: typeof permissionApi;
 	projectApi: typeof projectApi;
 	projectApiKeysApi: typeof projectApiKeysApi;
+	projectGroupApi: typeof projectGroupApi;
+	projectPermissionsApi: typeof projectPermissionsApi;
 	storage: Storage;
 	toastNotifier: ToastNotifier;
 	userApi: typeof userApi;