+++ apps/frontend/src/libs/modules/store/libs/types/root-reducer.type.ts
@@ -3,6 +3,8 @@ import { type reducer as contributorsReducer } from "~/modules/contributors/cont
 import { type reducer as groupsReducer } from "~/modules/groups/groups.js";
 import { type reducer as permissionsReducer } from "~/modules/permissions/permissions.js";
 import { type reducer as projectApiKeysReducer } from "~/modules/project-api-keys/project-api-keys.js";
+import { type reducer as projectGroupsReducer } from "~/modules/project-groups/project-groups.js";
+import { type reducer as projectPermissionsReduser } from "~/modules/project-permissions/project-permissions.js";
 import { type reducer as projectsReducer } from "~/modules/projects/projects.js";
 import { type reducer as usersReducer } from "~/modules/users/users.js";
 
@@ -12,6 +14,8 @@ type RootReducer = {
 	groups: ReturnType<typeof groupsReducer>;
 	permissions: ReturnType<typeof permissionsReducer>;
 	projectApiKeys: ReturnType<typeof projectApiKeysReducer>;
+	projectGroups: ReturnType<typeof projectGroupsReducer>;
+	projectPermissions: ReturnType<typeof projectPermissionsReduser>;
 	projects: ReturnType<typeof projectsReducer>;
 	users: ReturnType<typeof usersReducer>;
 };