+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -6,6 +6,7 @@ const APIPath = {
 	PERMISSIONS: "/permissions",
 	PROJECT_API_KEYS: "/project-api-keys",
 	PROJECT_GROUPS: "/project-groups",
+	PROJECT_PERMISSIONS: "/project-permissions",
 	PROJECTS: "/projects",
 	USERS: "/users",
 } as const;